Handlerfunktioenen ausgelagert
